Fix a few minor signedness warnings
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2488 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxExpectation.c
@@ -319,7 +319,7 @@ omxNewInternalExpectation(const char *expType, omxState* os)
 	omxExpectation* expect = Calloc(1, omxExpectation);
 
 	/* Switch based on Expectation type. */ 
-	for (int ex=0; ex < OMX_STATIC_ARRAY_SIZE(omxExpectationSymbolTable); ex++) {
+	for (size_t ex=0; ex < OMX_STATIC_ARRAY_SIZE(omxExpectationSymbolTable); ex++) {
 		const omxExpectationTableEntry *entry = omxExpectationSymbolTable + ex;
 		if(strncmp(expType, entry->name, MAX_STRING_LEN) == 0) {
 		        expect->expType = entry->name;
